Add search functionality and refactor Lol connection
