+++ app/assets/stylesheets/application.css.scss
@@ -137,6 +137,7 @@ input {
   .navbar-collapse{
     .search-form{
       margin-top:10px;
+      display: inline;
     }   
   }
 }
@@ -188,18 +189,65 @@ td {
 	vertical-align: top;
 }
 
+
+.header {
+  background-color: black;
+  color: white;
+  padding: 10px;
+  margin: 20px 0px 20px 10px;
+  border: 1px solid black;
+}
+.search-page {
+	position: fixed;
+	width: 100%;
+	height: 100%;
+	background-image: image-url("login_bg.png");
+	background-repeat: no-repeat;
+	background-size: cover;
+	padding-top: 20px;
+	.search-table{
+    border-collapse: collapse;
+    margin-top:60px;
+	  border: 1px solid black;
+	  cellpadding: 0;
+	  background: linear-gradient(#c8bbb6, #f4f2f1);
+	  .search-header{
+	    text-align: center;
+	    font-size: 1.7em;
+	  }
+	}
+	.search-table td, .search-table th{
+	  width: auto;
+	  padding: 5px;
+		border: 0px;
+		.btn{
+      width: 100%;
+    }
+    .btn-default{
+	    color: white;
+	    background: black;
+	  }
+	  .champ-img{
+      width: 10%;
+      vertical-align: middle;
+    }
+	  .no-stats{
+	    background: black;
+	    color: white;
+	    border: 2px solid black;
+	    border-radius: 5px;
+	  }
+
+	}
+}
+
+
+
+
 .spec-social {
 	width:100%;
 	height:100%;
-	.header {
-	  background-color: black;
-	  color: white;
-	  padding: 10px;
-	  margin-left:10px;
-	  margin-top:20px;
-	  margin-bottom:20px;
-	  border: 1px solid black;
-  }
+	
 
   .champions-header{
     margin-right: 20px;
@@ -238,3 +286,4 @@ td {
 	  }
   }
 }
+
