+++ app/views/layouts/_navigation.html.erb
@@ -33,15 +33,20 @@
             </ul>
         </li>
       </ul>
-      
-      <div class="col-md-3 search-form pull-right">
+    <div class="col-md-3 search-form pull-right">  
+      <%= form_tag search_summoner_path, class: "navbar-form", method: :get, id: 'summoner-lookup-form' do %>
         <div class="input-group">
-          <input type="text" class="form-control" placeholder="Enter Friend...">
+          <%= text_field_tag :summoner, params[:summoner], class: "form-control search-box", placeholder: "Search Summoner ", autofocus: true %>
           <span class="input-group-btn">
-            <button class="btn btn-default" type="submit">Find</button>
+            <%= button_tag(type: :submit, class: "btn btn-default") do %>Search
+            <% end %>
           </span>
-        </div><!-- /input-group -->
-      </div><!-- /.search-form -->
+        </div>
+      <% end %>
+    </div>
+          
+      
+      
 
     </div><!-- /.navbar-collapse -->
   </div><!-- /.container-fluid -->
