+++ app/views/layouts/application.html.erb
@@ -14,5 +14,7 @@
 		<% end %>
 
 	<%= yield %>
+
 </body>
+
 </html>
