+++ app/views/search/search.html.erb
@@ -1,0 +1,96 @@
+<div class = "search-page" id="summoner-lookup">
+  
+  <div class = "col-md-5 col-md-offset-3">
+  	<div class="table-responsive">
+		  <table class="table search-table" cellspacing="0" style="border-spacing: 0;">
+		    <thead>
+		    	<tr>
+		    		<th colspan="3" scope="colgroup" class="header search-header">Summoner Search Result</th>
+		    	</tr>
+		    </thead>
+		    <tbody>
+		      <% if @summoner %>
+		        <tr>
+		        	<td ><strong>Summoner Name: <%= @name.capitalize %></strong></td>
+              <% if @statsObj %>
+		        	  <td ><strong>Most Played Champ: <%= "#{@mpChamp[:name]}"%> </strong><%= " - (#{@mostPlayedChampObj.stats.total_sessions_played}x)"%></td>
+		        	  <td class="champ-img" rowspan="3" scope="rowgroup"><%= link_to image_tag("icons/#{@mpChamp[:name].delete(' ')}Square.png"), champion_path(@mpChamp.id)  %></td>
+		        	<% else %>
+                <td colspan="2" scope="colgroup" ><strong>No stats found for summoner.</strong></td>
+              <% end %>
+		        </tr>
+		        <tr>
+		        	<td ><strong>Summoner Level: <%= @summoner.summoner_level %></strong></td>
+		        	<% if @statsObj %>
+			        	<td ><strong>Highest KDR Champ: <%= "#{@hkdrChamp[:name]}"%> </strong> 
+				    		                                    <% if @highestKillDeathRatioChampObj[0].stats.total_deaths_per_session >  0 %>
+		                                                  <%= " - (#{((@highestKillDeathRatioChampObj[0].stats.most_champion_kills_per_session.to_f)/(@highestKillDeathRatioChampObj[0].stats.total_deaths_per_session)).round(2)})"%>
+		                                                <%else%>
+		                                                  <%= " - (#{(@highestKillDeathRatioChampObj[0].stats.most_champion_kills_per_session_f.round(2))})"%>
+		                                                <%end%>
+			        	</td>
+			        <% else %>
+			          <td colspan="2" scope="colgroup" rowspan="2" scope="rowgroup"><strong>Either summoner is inactive or there are no stats to display.</strong></td>
+			        <% end %>
+		        </tr>
+		        <tr>
+		          <td ><strong>Last Active: <%= @summoner.revision_date.to_s.slice(0,10) %></strong></td>
+		          <% if @statsObj %>
+		            <td ><strong>Highest WLR Champ: <%= "#{@hwlrChamp[:name]}"%> </strong> 
+			    		                                    <% if @highestWinLossRatioChampObj[0].stats.total_deaths_per_session >  0 %>
+	                                                  <%= " - (#{((@highestWinLossRatioChampObj[0].stats.most_champion_kills_per_session.to_f)/(@highestWinLossRatioChampObj[0].stats.total_deaths_per_session)).round(2)})"%>
+	                                                <%else%>
+	                                                  <%= " - (#{(@highestWinLossRatioChampObj[0].stats.most_champion_kills_per_session_f.round(2))})"%>
+	                                                <%end%>
+	              </td>
+		          <% end %>
+		        </tr>
+		      <% else %>
+		        <tr>
+		        	<td align="center" colspan="3" scope="colgroup" ><strong><%= @name.capitalize %> is not a recognized League of Legends player. </strong>
+		        	</td>
+		        </tr>
+		        <tr>
+		        	<td align="center" colspan="3" scope="colgroup" rowspan="2" scope = "rowgroup"><strong> Verify that <%= @name.capitalize %> is a valid summoner name and try again.</strong>
+		        	</td>
+		        </tr>
+    		  <% end %>
+		    </tbody>
+		    <tfoot>
+		      <tr>
+            <%if logged_in? %>		      
+			      	<% if @statsObj %>
+                <% @user = User.find_summoner_by_name(@name) %>
+                <% if @user %><!--more logic needed here to confirm that user is not already a friend-->
+                  <td><%= link_to "Add Summoner".html_safe, profile_path(@current_user.id), :class =>"btn btn-success" %></td>
+                  <td ><%= link_to "View Profile".html_safe, profile_path(@user.id), :class =>"btn btn-primary" %></td>
+                  <td ><%= link_to "Dismiss Search".html_safe, profile_path(@current_user.id), :class =>"btn btn-danger" %></td>
+                <% else %>
+                  <td ><%= link_to "Invite Friend".html_safe, root_path, :class =>"btn btn-primary" %></td>
+                  <td   colspan="2" scope="colgroup"><%= link_to "Dismiss Search".html_safe, profile_path(@current_user.id), :class =>"btn btn-danger" %></td>
+                <% end %>
+			      	<% else %>
+			      	  <td   colspan="3" scope="colgroup"><%= link_to "Dismiss Search".html_safe, profile_path(@current_user.id), :class =>"btn btn-danger" %></td>
+			      	<% end %>
+			      <% else %>
+			        <% if @statsObj %>
+                <% @user = User.find_summoner_by_name(@name) %>
+                 <% if @user %>
+                  <td  ><%= link_to "Login to Add Summoner".html_safe, login_path, :class =>"btn btn-success" %></td>
+                  <td ><%= link_to "Sign Up to Add Summoner".html_safe, new_profile_path, :class =>"btn btn-primary" %></td>
+                 <% else %>
+                  <td   ><%= link_to "Login to Invite Summoner".html_safe, login_path, :class =>"btn btn-success" %></td>
+                  <td  ><%= link_to "Sign Up and Invite Summoner".html_safe, new_profile_path, :class =>"btn btn-primary" %></td>
+                 <% end %>
+                <td ><%= link_to "Dismiss Search".html_safe, root_path, :class =>"btn btn-danger" %></td>
+              <% else %>
+                <td  colspan="3" scope="colgroup"><%= link_to "Dismiss Search".html_safe, root_path, :class =>"btn btn-danger" %></td>
+              <% end %>
+			      <% end %>
+		      </tr>
+		    </tfoot>
+		  </table>
+    </div>
+  </div>
+</div>
+
