+++ config/routes.rb
@@ -18,6 +18,5 @@ Rails.application.routes.draw do
   resources :champions, only: [:show]
   resources :account_activations, only: [:edit]
   resources :password_resets,     only: [:new, :create, :edit, :update]
-  
-
+  get 'search_summoner', to: 'search#search'
 end
